Extract NoEntityView component that can be used for multiple pages
